requirements_dev.txt
opengl-registry
python-dotenv
